Code Refactoring. JIRA-SHAR-55: Numeric mode for ETB + KES shouldn't have decimal places
